migrate to prepClipOperations
